+++ src/policy/core.rs
@@ -1,8 +1,9 @@
-use dom_query::{Document, NodeRef};
+use dom_query::NodeRef;
 use html5ever::LocalName;
 use tendril::StrTendril;
 
 use super::builder::PolicyBuilder;
+use crate::macros::sanitize_methods;
 use crate::traits::{SanitizeDirective, SanitizePolicy};
 use crate::{Permissive, Restrictive};
 
@@ -37,24 +38,7 @@ pub struct Policy<'a, T: SanitizeDirective = Restrictive> {
 }
 
 impl<T: SanitizeDirective> Policy<'_, T> {
-    /// Sanitizes a node by applying the policy rules according to the directive type.
-    ///
-    /// For [Permissive] directive: Removes elements and attributes specified in the policy.
-    /// For [Restrictive] directive: Keeps only elements and attributes specified in the policy.
-    pub fn sanitize_node(&self, node: &dom_query::NodeRef) {
-        T::sanitize_node(self, node);
-        node.normalize();
-    }
-    /// Sanitizes the document.
-    pub fn sanitize_document(&self, document: &Document) {
-        self.sanitize_node(&document.root());
-    }
-    /// Sanitizes the HTML content by applying the policy rules according to the directive type.
-    pub fn sanitize_html<S: Into<StrTendril>>(&self, html: S) -> StrTendril {
-        let doc = Document::from(html);
-        self.sanitize_document(&doc);
-        doc.html()
-    }
+    sanitize_methods!();
 }
 
 impl<T: SanitizeDirective> SanitizePolicy for Policy<'_, T> {